moved responses to their own file and added boilerplate for count_response
