+++ src/count_response.py
@@ -1,0 +1,8 @@
+from telegram import Update
+from telegram.ext import ContextTypes
+
+# Placeholder, will use db
+counter = 0
+
+async def count_response(update: Update, _context: ContextTypes.DEFAULT_TYPE):
+    ...
